+++ app/controllers/home_controller.rb
@@ -21,7 +21,9 @@ class HomeController < ApplicationController
       { url: about_url, lastmod: view_mtime("about") },
       { url: privacy_url, lastmod: view_mtime("privacy") },
       { url: cookies_url, lastmod: view_mtime("cookies") },
-      { url: terms_url, lastmod: view_mtime("terms") }
+      { url: terms_url, lastmod: view_mtime("terms") },
+      { url: root_url + "llms.txt", lastmod: public_file_mtime("llms.txt") },
+      { url: root_url + "llms-full.txt", lastmod: public_file_mtime("llms-full.txt") }
     ]
 
     render layout: false
@@ -31,5 +33,9 @@ class HomeController < ApplicationController
 
   def view_mtime(name)
     Rails.root.join("app/views/home/#{name}.html.erb").mtime.to_date.iso8601
+  end
+
+  def public_file_mtime(name)
+    Rails.root.join("public/#{name}").mtime.to_date.iso8601
   end
 end